refactor: 패키지 및 컨벤션 수정
