+++ src/main/java/ladder/domain/generator/LineGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.generator;
 
 public interface LineGenerator {
     boolean generate();
