+++ src/test/java/ladder/domain/UsersTest.java
@@ -1,6 +1,5 @@
 package ladder.domain;
 
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
